Python-Backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
bcrypt==3.2.0
beanie==2.0.1
bidict==0.23.1
cffi==2.0.0
click==8.3.1
colorama==0.4.6
dnspython==2.8.0
exceptiongroup==1.3.1
fastapi==0.129.0
h11==0.16.0
idna==3.11
lazy-model==0.4.0
motor==3.7.1
passlib==1.7.4
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
PyJWT==2.11.0
pymongo==4.16.0
python-dotenv==1.2.1
python-engineio==4.13.1
python-socketio==5.16.1
simple-websocket==1.1.0
six==1.17.0
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.40.0
wsproto==1.3.2
google-auth==2.38.0
requests==2.32.3
firebase-admin==6.6.0
